requirements.txt
# PCO API Wrapper - Python Dependencies

# Core dependencies
pypco>=1.2.0              # Planning Center Online API wrapper
flask>=2.3.0              # Web framework for REST API
python-dotenv>=1.0.0      # Environment variable management
requests>=2.31.0          # HTTP library (pypco dependency)
psycopg2-binary>=2.9.9    # PostgreSQL database adapter
# Caching dependencies (optional)
redis>=5.0.0              # Redis cache backend (optional)


# Development dependencies (optional)
pytest>=7.4.0             # Testing framework
pytest-cov>=4.1.0         # Test coverage
black>=23.7.0             # Code formatter
flake8>=6.1.0             # Linter
mypy>=1.5.0               # Type checker